rename shared stuff
